backend/requirements.txt
# Web framework
fastapi==0.117.1
pydantic==2.4.1
uvicorn==0.32.0

# Data processing
pandas==2.3.2
scikit_learn==1.6.1
prophet==1.1.7
matplotlib==3.8.0

# Networking
requests==2.32.5
aiohttp==3.9.5
httpx==0.27.2

# ML / Search
duckduckgo_search==8.1.1

# Utilities
nest_asyncio==1.6.0
python-dotenv==1.0.1
apscheduler==3.10.4

# Browser automation
selenium==4.25.0

# Database / API client
supabase==2.9.1

# Telegram bot
aiogram==3.3.0

# Chromium Driver
chromedriver-binary==117.0.5938.92

# Google Generative AI
google-generativeai==0.8.5

# Прочие утилиты
urllib3
plotly
